sal: try to avoid abuse of OUStringBuffer(int) ctor
... to avoid bugs like commit f0d6e0e1e21afd0adf5bd01d771b2d83d8f13a48.

+++ xmloff/source/text/txtparai.cxx
@@ -133,7 +133,7 @@ void XMLCharContext::EndElement()
         }
         else
         {
-            OUStringBuffer sBuff( m_nCount );
+            OUStringBuffer sBuff(static_cast<int>(m_nCount));
             while( m_nCount-- )
                 sBuff.append( &m_c, 1 );
